separate `AstVisitor` from `QueryAst.scala`
